Introduce the `MandrillEmail` to prevent orphans.
This commit introduces a newtype wrapper called `MandrillEmail` to wrap
the internal `EmailAddress` and thus avoid the need for a
ToJSON/FromJSON orphan instances.

Despite the newtype introduction the API should not break for the end
user, provided he created the `MandrillRecipient` via the smart
constructor, and he's not poking around with the `MandrillMessage`. We
should probably avoid exporting the entire Types.hs module, but only
types & smart constructors (and lenses!) but that is out of the scope of
this PR.

As low hanging fruit, I have ported the project to stack.

@gregwebs Are you cool with me merging? I will release the new Mandrill
upstream on Hackage soon afterwards.